azure-storage/requirements.txt
python-dotenv
requests
azure-storage-blob==12.3.2